Added updated String recommended implementation and added unit tests
